Move some specimen comment actions
